[WIP] setup: keep honoring selected extensions even in v0 repository
Earlier we tried to correct the way extensions.* are handled in a
repository whose repositoryFormatVersion is v0---any extensions
known to the version of Git that happened to be running were honored
by mistake, instead of getting ignored.

The proposed correction was to ignore them and also refuse to
upgrade the repositoryFormatVersion to v1 when any extensions.*
exist in v0 repository.  This is a safe approach (the user may have
set a variable whose name happens to begin with "extensions.",
knowing that it would not interfere with what Git would as long as
the repository is v0 where extensions are ignored), but is a bit too
strict.  Those extensions that used to be honored by mistake in v0
repositories couldn't have been used for any other purpose in their
repository, so there is no reason not to allow upgrading to v1 to
honor them by design and not by mistake.

With this change, we declare that v0 and v1 are largely the same
with respect to a few selected extensions that used to be honored by
mistake anyway, and allow upgrading v0 to v1 as long as no other
"extensions.*" configuration variables exist.

Signed-off-by: Junio C Hamano <gitster@pobox.com>