Read FETCH_HEAD as loose ref
FETCH_HEAD is a loose ref (either symref or OID), followed by further
metadata. It can therefore not be read through a ref backend
normally. Special case this in reftable-backend.c.

This functionality is shared between all backends, but the ref backend
interface doesn't have a method to retrieve the repository directory,
hence this functionality must be pushed down to the backend
implementation.

Signed-off-by: Han-Wen Nienhuys <hanwen@google.com>