diff: teach --stat to ignore uninteresting modifications
Sometimes when diffing, files may show as being momdified even when
there are no interesting diffs to show.  This happens naturally when
using options such as --ignore-space-change.  We don't want to prevent
the display  of all files that have 0 effective diffs since they could
be the result of a rename, permission change, or other similar operation
that may still be of interest so we special case additions and deletions
as they are always interesting.

Signed-off-by: Matthew Rogers <mattr94@gmail.com>