fetch: do not look for submodule changes in unchanged refs
This operation is very expensive, as it scans all the refs using
setup_revisions, which resolves each ref, including checking if it
is ambiguous, or if it is a file name etc.

There is no reason to do all that for refs that haven't changed in this
fetch.

Reported here:
https://public-inbox.org/git/CAGHpTBKSUJzFSWc=uznSu2zB33qCSmKXM-iAjxRCpqNK5bnhRg@mail.gmail.com/

Amends commit be76c2128234d94b47f7087152ee55d08bb65d88.

Signed-off-by: Orgad Shaneh <orgads@gmail.com>