submodule: suppress checking for file name ambiguity for object ids
The argv argument of collect_changed_submodules() contains obly object ids
(submodule references).

Notify setup_revisions() that the input is not filenames by passing
assume_dashdash, so it can avoid redundant stat for each ref.

A better improvement would be to pass oid_array instead of stringified argv,
but that will require a larger change, which can be done later.

Signed-off-by: Orgad Shaneh <orgads@gmail.com>